% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{build_find_taxon}
\alias{build_find_taxon}
\title{Find list of unique datasets within compilation containing specified taxa}
\usage{
build_find_taxon(taxa, austraits, original_name = FALSE)
}
\arguments{
\item{austraits}{AusTraits compilation}

\item{original_name}{Logical; if TRUE use column in compilation which contains original
species names, default = FALSE}

\item{taxon_name}{A vector which contains species names}
}
\value{
List of unique datasets within compilation containing each taxon
}
\description{
Find list of unique datasets within compilation containing specified taxa
}
