% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{check_pivot_duplicates}
\alias{check_pivot_duplicates}
\title{Identify duplicates preventing pivoting wider}
\usage{
check_pivot_duplicates(
  database_object,
  dataset_ids = unique(database_object$traits$dataset_id)
)
}
\arguments{
\item{database_object}{Database object}

\item{dataset_ids}{\code{dataset_id}'s to check for duplicates; default is all of them}
}
\value{
Tibble with duplicates and pivot columns
}
\description{
Identify duplicates preventing pivoting wider
}
