% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{check_pivot_wider}
\alias{check_pivot_wider}
\title{Test whether a dataset can pivot wider}
\usage{
check_pivot_wider(dataset)
}
\arguments{
\item{dataset}{Built dataset with \code{test_build_dataset}}
}
\value{
Number of rows with duplicates preventing pivoting wider
}
\description{
Test whether the traits table of a dataset can pivot wider with the minimum required columns.
}
