% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_df_to_list}
\alias{convert_df_to_list}
\title{Convert dataframe to list}
\usage{
convert_df_to_list(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A (yaml) list
}
\description{
Convert a dataframe to a named list, useful when converting to yaml.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
convert_df_to_list(dplyr::starwars)
}
