% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_tree_branch}
\alias{create_tree_branch}
\title{Format a tree structure from a vector}
\usage{
create_tree_branch(x, title, prefix = "")
}
\arguments{
\item{x}{Vector of terms}

\item{title}{Name of branch}

\item{prefix}{Specifies the amount of indentation}
}
\value{
Vector of character strings for the tree structure
}
\description{
`create_tree_branch()` is used to create a tree structure to show how things
are related. In AusTraits, this is used in the vignettes to show the file
structure of the repository and also to show the different components of the
AusTraits database.
}
