% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{dataset_configure}
\alias{dataset_configure}
\title{Configure AusTraits database object}
\usage{
dataset_configure(filename_metadata, definitions)
}
\arguments{
\item{filename_metadata}{Metadata yaml file for a given study}

\item{definitions}{Definitions read in from the \code{traits.yml}}
}
\value{
List with \code{dataset_id}, \code{metadata}, \code{definitions} and \code{unit_conversion_functions}
}
\description{
Creates the config object which gets passed onto \code{dataset_process}. The config list contains
the subset of definitions and unit conversions for those traits for a each study.
\code{dataset_configure} is used in the \code{remake::make} process to configure individual studies mapping the
individual traits found in that study along with any relevant unit conversions
and definitions. \code{dataset_configure} and \code{dataset_process} are applied to every study
in the \code{remake.yml} file.
}
\examples{
\dontrun{
dataset_configure("data/Falster_2003/metadata.yml", read_yaml("config/traits.yml"))
}
}
