% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{dataset_process}
\alias{dataset_process}
\title{Load Dataset}
\usage{
dataset_process(
  filename_data_raw,
  config_for_dataset,
  schema,
  resource_metadata,
  unit_conversion_functions,
  filter_missing_values = TRUE
)
}
\arguments{
\item{filename_data_raw}{Raw `data.csv` file for any given study}

\item{config_for_dataset}{Config settings generated from `dataset_configure()`}

\item{schema}{Schema for traits.build}

\item{resource_metadata}{Metadata about the traits compilation read in from the config folder}

\item{unit_conversion_functions}{`unit_conversion.csv` file read in from the config folder}

\item{filter_missing_values}{Default filters missing values from the excluded data table;
change to false to see the rows with missing values.}
}
\value{
List, AusTraits database object
}
\description{
`dataset_process` is used to load individual studies using the config file generated
from `dataset_configure()`. `dataset_configure` and `dataset_process` are applied to every
study in the `remake.yml` file.
}
\examples{
\dontrun{
dataset_process("data/Falster_2003/data.csv", dataset_configure("data/Falster_2003/metadata.yml",
read_yaml("config/traits.yml")),
get_schema(),
get_schema("config/metadata.yml", "metadata"),
get_unit_conversions("config/unit_conversions.csv"))
}
}
