% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{dataset_test}
\alias{dataset_test}
\title{Test whether specified \code{dataset_id} has the correct setup}
\usage{
dataset_test(
  dataset_ids,
  path_config = "config",
  path_data = "data",
  reporter = testthat::CompactProgressReporter
)
}
\arguments{
\item{dataset_ids}{Vector of \code{dataset_id} for sources to be tested}

\item{path_config}{Path to folder containing configuration files}

\item{path_data}{Path to folder containing data files}

\item{reporter}{\code{testthat} reporter to use to summarise output}
}
\description{
Run tests to ensure that specified \code{dataset_id} has the correct setup.
}
