% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{dataset_update_taxonomy}
\alias{dataset_update_taxonomy}
\title{Apply taxonomic updates to austraits_raw}
\usage{
dataset_update_taxonomy(austraits_raw, taxa)
}
\arguments{
\item{austraits_raw}{AusTraits compiled data as a large list without taxonomic updates applied}

\item{taxa}{Taxon list}
}
\value{
List of AusTraits compiled data with taxonomic updates applied
}
\description{
Applies taxonomic updates to austraits_raw.
}
