% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_schema}
\alias{get_schema}
\title{Load schema for an traits.build data compilation (excluding traits)}
\usage{
get_schema(
  path = system.file("support", "traits.build_schema.yml", package = "traits.build"),
  subsection = NULL
)
}
\arguments{
\item{path}{path to schema file. By default loads version included with the package}

\item{subsection}{section to load}
}
\value{
a list
}
\description{
Load schema for an traits.build data compilation (excluding traits)
}
\examples{
{

schema <- get_schema()
}
}
