% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_add_identifiers}
\alias{metadata_add_identifiers}
\title{For specified \code{dataset_id} import database cross-referencing identifer data from a dataframe}
\usage{
metadata_add_identifiers(dataset_id, overwrite = FALSE)
}
\arguments{
\item{dataset_id}{Identifier for a particular study in the database}

\item{overwrite}{Overwrite existing information}
}
\description{
This functions asks users which columns in the dataframe they would like to keep
and records this appropriately in the metadata. The input data is assumed to be
in wide format.
The output may require additional manual editing if the user selects a non-standard identifier type.
}
