% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_add_source_bibtex}
\alias{metadata_add_source_bibtex}
\title{Adds citation details to a metadata file for given study}
\usage{
metadata_add_source_bibtex(
  dataset_id,
  file,
  type = "primary",
  drop = c("dateobj", "month")
)
}
\arguments{
\item{dataset_id}{Identifier for a particular study in the database}

\item{file}{Name of file where reference is saved}

\item{type}{Type of reference: \code{primary}, \code{secondary} or \code{original}
(or \code{original_01}, \code{original_02}, etc., for multiple sources)}

\item{drop}{Variables in bibtex to ignore}
}
\value{
\code{metadata.yml} file with citation details added
}
\description{
Adds citation details to a metadata file for given study
}
