% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_add_taxonomic_change}
\alias{metadata_add_taxonomic_change}
\title{Add a taxonomic change into the `metadata.yml` file for a `dataset_id`}
\usage{
metadata_add_taxonomic_change(
  dataset_id,
  find,
  replace,
  reason,
  taxonomic_resolution
)
}
\arguments{
\item{dataset_id}{Identifier for a particular study in the database}

\item{find}{Original name used by the contributor}

\item{replace}{Taxonomic name accepted by APC or APNI}

\item{reason}{Reason for taxonomic change}

\item{taxonomic_resolution}{The rank of the most specific taxon name (or scientific name)
to which a submitted orignal name resolves}
}
\value{
`metadata.yml` file with taxonomic change added
}
\description{
Add a single taxonomic change into the `metadata.yml` file for a specific study.
}
