% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_add_taxonomic_changes_list}
\alias{metadata_add_taxonomic_changes_list}
\title{Add a list of taxonomic updates into a metadata file for a \code{dataset_id}}
\usage{
metadata_add_taxonomic_changes_list(dataset_id, taxonomic_updates)
}
\arguments{
\item{dataset_id}{Identifier for a particular study in the database}

\item{taxonomic_updates}{Dataframe of taxonomic updates}
}
\value{
\code{metadata.yml} file with multiple taxonomic updates added
}
\description{
Add multiple taxonomic changes to the \code{metadata.yml} file using a dataframe
containing the taxonomic changes to be made.
}
