% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_create_template}
\alias{metadata_create_template}
\title{Create a template of file \code{metadata.yml} for specified \code{dataset_id}}
\usage{
metadata_create_template(
  dataset_id,
  path = file.path("data", dataset_id),
  skip_manual = FALSE,
  user_responses = NULL
)
}
\arguments{
\item{dataset_id}{Identifier for a particular study in the database}

\item{path}{Location of file where output is saved}

\item{skip_manual}{Allows skipping of manual selection of variables, default = FALSE}

\item{user_responses}{Named list containing simulated user input for manual selection
of variables, mainly for testing purposes}
}
\value{
A yaml file template for metadata
}
\description{
Includes place-holders for major sections of the metadata.
}
