% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_exclude_observations}
\alias{metadata_exclude_observations}
\title{Exclude observations in a yaml file for a `dataset_id`}
\usage{
metadata_exclude_observations(dataset_id, variable, find, reason)
}
\arguments{
\item{dataset_id}{Identifier for a particular study in the database}

\item{variable}{Variable name}

\item{find}{Term to find by}

\item{reason}{Reason for exclusion}
}
\value{
`metadata.yml` file with excluded observations
}
\description{
Exclude observations in a yaml file for a `dataset_id`
}
