% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_find_taxonomic_change}
\alias{metadata_find_taxonomic_change}
\title{Find \code{dataset_id}'s with a given taxonomic change}
\usage{
metadata_find_taxonomic_change(find, replace = NULL, studies = NULL)
}
\arguments{
\item{find}{Name of original species}

\item{replace}{Name of replacement species, default = NULL}

\item{studies}{Name of studies to look through, default = NULL}
}
\description{
Find \code{dataset_id}'s with a given taxonomic change
}
