% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_remove_taxonomic_change}
\alias{metadata_remove_taxonomic_change}
\title{Remove a taxonomic change from a yaml file for a \code{dataset_id}}
\usage{
metadata_remove_taxonomic_change(dataset_id, find)
}
\arguments{
\item{dataset_id}{Identifier for a particular study in the database}

\item{find}{Taxonomic name to find}
}
\value{
\code{metadata.yml} file with a taxonomic change removed
}
\description{
Remove a taxonomic change from a yaml file for a \code{dataset_id}
}
