% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_update_taxonomic_change}
\alias{metadata_update_taxonomic_change}
\title{Update a taxonomic change into a yaml file for a `dataset_id`}
\usage{
metadata_update_taxonomic_change(
  dataset_id,
  find,
  replace,
  reason,
  taxonomic_resolution
)
}
\arguments{
\item{dataset_id}{Identifier for a particular study in the database}

\item{find}{Original taxonomic name}

\item{replace}{Updated taxonomic name to replace original taxonomic name}

\item{reason}{Reason for change}

\item{taxonomic_resolution}{The rank of the most specific taxon name (or scientific name)
to which a submitted orignal name resolves}
}
\value{
`metadata.yml` file with added substitution
}
\description{
Update a taxonomic change into a yaml file for a `dataset_id`
}
