% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_user_select_column}
\alias{metadata_user_select_column}
\title{Select column by user}
\usage{
metadata_user_select_column(column, choices)
}
\arguments{
\item{column}{Name of the variable of interest}

\item{choices}{The options that can be selected from}
}
\description{
\code{metadata_user_select_column} is used to select which columns in a dataframe/tibble
corresponds to the variable of interest. It is used to compile the metadata yaml
file by prompting the user to choose the relevant columns. It is used in
\code{metadata_add_locations} and \code{metadata_create_template}.
}
