% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_user_select_names}
\alias{metadata_user_select_names}
\title{Select variable names by user}
\usage{
metadata_user_select_names(title, vars)
}
\arguments{
\item{title}{Character string providing the instruction for the user}

\item{vars}{Variable names}
}
\description{
\code{metadata_user_select_names} is used to prompt the user to select the variables that
are relevant for compiling the metadata yaml file. It is currently used for
\code{metadata_add_traits}, \code{metadata_add_locations} and \code{metadata_add_contexts}.
}
