% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{new_taxa_trait_combinations}
\alias{new_taxa_trait_combinations}
\title{New taxa added by a dataset}
\usage{
new_taxa_trait_combinations(database, dataset)
}
\arguments{
\item{database}{traits.build database}

\item{dataset}{dataset_id for dataset of interest}
}
\value{
table with number of new taxa added for each trait
}
\description{
Function to indicate how many new taxa by trait combinations are added with a dataset.
}
\examples{
\dontrun{
new_taxa_trait_combinations(austraits, "Falster_2003")
}
}
