% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_convert_units}
\alias{process_convert_units}
\title{Convert units to desired type}
\usage{
process_convert_units(data, definitions, unit_conversion_functions)
}
\arguments{
\item{data}{Tibble or dataframe containing the study data}

\item{definitions}{Definitions read in from the \code{traits.yml} file in the config folder}

\item{unit_conversion_functions}{\code{unit_conversions.csv} file stored in the config folder}
}
\value{
Tibble with converted units
}
\description{
Convert units to desired type
}
