% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_flag_unsupported_values}
\alias{process_flag_unsupported_values}
\title{Flag disallowed trait values and disallowed characters}
\usage{
process_flag_unsupported_values(data, definitions)
}
\arguments{
\item{data}{Tibble or dataframe containing the study data}

\item{definitions}{Definitions read in from the `traits.yml` file in the config folder}
}
\value{
Tibble with flagged values that are unsupported categorical trait values, missing values
or numeric trait values that cannot be converted to numeric
}
\description{
Flags any categorical traits values that are not on the list of allowed values defined in the
`traits.yml` file.
NA values are flagged as errors.
Numeric values that cannot convert to numeric are also flagged as errors.
}
