% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_format_contributors}
\alias{process_format_contributors}
\title{Format contributors from list into tibble}
\usage{
process_format_contributors(my_list, dataset_id, schema)
}
\arguments{
\item{my_list}{List of input information}

\item{dataset_id}{Identifier for a particular study in the AusTraits database}

\item{schema}{Schema for traits.build}
}
\value{
Tibble with details of contributors
}
\description{
Format contributors, read in from the \code{metadata.yml} file. Converts from list to tibble.
}
\examples{
\dontrun{
process_format_contributors(read_metadata("data/Falster_2003/metadata.yml")$contributors)
}
}
