% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_generate_id}
\alias{process_generate_id}
\title{Function to generate sequence of integer ids from vector of names
Determines number of 00s needed based on number of records.}
\usage{
process_generate_id(x, prefix, sort = FALSE)
}
\arguments{
\item{x}{Vector of text to convert}

\item{prefix}{Text to put before id integer}

\item{sort}{Logical to indicate whether x should be sorted before ids are generated}
}
\value{
Vector of ids
}
\description{
Function to generate sequence of integer ids from vector of names
Determines number of 00s needed based on number of records.
}
