% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_generate_method_ids}
\alias{process_generate_method_ids}
\title{Function to generate sequence of integer ids for methods}
\usage{
process_generate_method_ids(metadata_traits)
}
\arguments{
\item{metadata_traits}{the traits section of the metadata}
}
\value{
Tibble with traits, methods, and method_id
}
\description{
Function to generate sequence of integer ids for methods
}
