% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_csv_char}
\alias{read_csv_char}
\title{Read in a csv as a tibble with column types as characters}
\usage{
read_csv_char(...)
}
\arguments{
\item{...}{Arguments passed to the \code{read_csv()}}
}
\value{
A tibble
}
\description{
Reads in a csv file using the \code{read_csv} function from readr
with columns as characters.
}
