% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_metadata}
\alias{read_metadata}
\title{Read in a \code{metadata.yml} file for a study}
\usage{
read_metadata(path)
}
\arguments{
\item{path}{Location of the metadata file}
}
\description{
A dataset's \verb{locations:} may either list the locations inline, or name a csv
file beside \code{metadata.yml} -- \code{locations: locations.csv} -- which is read in
its place and returned as a tibble.
}
