% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{trait_pivot_wider}
\alias{trait_pivot_wider}
\title{Pivot long format data into a wide format}
\usage{
trait_pivot_wider(traits)
}
\arguments{
\item{traits}{The traits table from database (list object)}
}
\value{
A tibble in wide format
}
\description{
`trait_pivot_wider` "widens" long format data ("tidy data").

Databases built with `traits.build` are organised in a long format where observations are on different rows and the
type of observation is denoted by various identifying columns (e.g `trait_name`, `dataset_id`,
`observation_id`, etc.).
This function converts the data into wide format so that each trait in its own column.
}
\details{
`trait_pivot_wider` will return a single wide tibble; note that some meta-data columns
(unit, replicates, measurement_remarks, basis_of_record, basis_of_value) will be excluded to
produce a useful wide tibble.
}
\examples{
\dontrun{
data <- austraits$traits \%>\% filter(dataset_id == "Falster_2003")
data # Long format
traits_wide <- trait_pivot_wider(data)
traits_wide # Wide format
}
}
\author{
Daniel Falster - daniel.falster@unsw.edu.au
}
