% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{util_check_disallowed_chars}
\alias{util_check_disallowed_chars}
\title{Check values in a vector do not contain disallowed characters}
\usage{
util_check_disallowed_chars(object)
}
\arguments{
\item{object}{Vector}
}
\value{
Vector of logical values
}
\description{
\code{util_check_disallowed_chars} checks if values in a vector do not contain disallowed characters,
i.e. values outside of ASCII.
}
