% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_list_to_df1}
\alias{util_list_to_df1}
\title{Convert a list with single entries to dataframe}
\usage{
util_list_to_df1(my_list)
}
\arguments{
\item{my_list}{A list with single entries}
}
\value{
A tibble with two columns
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontrun{
util_list_to_df1(as.list(dplyr::starwars)[2])
}
}
