% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_replace_null}
\alias{util_replace_null}
\title{Convert NULL values to a different value}
\usage{
util_replace_null(x, val = NA)
}
\arguments{
\item{x}{A NULL value or a non-NULL object}

\item{val}{Specify what the null value should be returned as, default is NA}
}
\value{
NA or a non-NULL object
}
\description{
\code{util_replace_null} converts NULL values a different value. Default is
converting NULL to NA.
}
\examples{
\dontrun{
util_replace_null(NULL)
}
}
