% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_chars.R
\name{util_tally_chars}
\alias{util_tally_chars}
\title{Tally the distinct characters in a report, most frequent first}
\usage{
util_tally_chars(report)
}
\arguments{
\item{report}{Report data frame}
}
\value{
A length-1 string
}
\description{
Tally the distinct characters in a report, most frequent first
}
\keyword{internal}
