% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_chars.R
\name{util_write_lines_utf8}
\alias{util_write_lines_utf8}
\title{Write lines to a file as UTF-8}
\usage{
util_write_lines_utf8(lines, path)
}
\arguments{
\item{lines}{Character vector to write}

\item{path}{File to write to}
}
\value{
Nothing, called for its side effect
}
\description{
Write lines to a file as UTF-8
}
\keyword{internal}
