% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{build_setup_pipeline}
\alias{build_setup_pipeline}
\title{Update the `remake.yml` file with new studies}
\usage{
build_setup_pipeline(
  template = readLines(system.file("support", "remake.yml.whisker", package =
    "traits.build")),
  path = "data",
  dataset_ids = dir(path)
)
}
\arguments{
\item{template}{Template used to build}

\item{path}{Path to folder with data}

\item{dataset_ids}{`dataset_id`'s to include; by default includes all}
}
\value{
Updated `remake.yml` file
}
\description{
`build_setup_pipeline` rewrites the `remake.yml` file to include new
studies.
}
