% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{build_setup_pipeline}
\alias{build_setup_pipeline}
\title{Update the \code{remake.yml} file with new studies}
\usage{
build_setup_pipeline(
  dataset_ids = dir("data"),
  method = "base",
  database_name = "database",
  template = select_pipeline_template(method),
  workers = 1
)
}
\arguments{
\item{dataset_ids}{\code{dataset_id}'s to include; by default includes all}

\item{method}{Approach to use in build}

\item{database_name}{Name of database to be built}

\item{template}{Template used to build}

\item{workers}{Number of workers/parallel processes to use when using
method = "furrr"}
}
\value{
Updated \code{remake.yml} file
}
\description{
\code{build_setup_pipeline} rewrites the \code{remake.yml} file to include new
studies.
}
