% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{build_setup_pipeline}
\alias{build_setup_pipeline}
\title{Update the `remake.yml` file with new studies}
\usage{
build_setup_pipeline(
  dataset_ids = dir("data"),
  method = "base",
  template = select_pipeline_template(method)
)
}
\arguments{
\item{dataset_ids}{`dataset_id`'s to include; by default includes all}

\item{method}{Approach to use in build}

\item{template}{Template used to build}
}
\value{
Updated `remake.yml` file
}
\description{
`build_setup_pipeline` rewrites the `remake.yml` file to include new
studies.
}
