% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_add_substitution}
\alias{metadata_add_substitution}
\title{Add a categorical trait value substitution into a metadata file for a `dataset_id`}
\usage{
metadata_add_substitution(dataset_id, trait_name, find, replace)
}
\arguments{
\item{dataset_id}{Identifier for a particular study in the database}

\item{trait_name}{The database defined name for a particular trait}

\item{find}{Trait value in the original data.csv file}

\item{replace}{Trait value supported by database}
}
\value{
`metadata.yml` file with a substitution added
}
\description{
`metadata_add_substitution` is used to align the categorical trait values used
by a contributor to the categorical values supported by the database. These values
are defined in the `traits.yml` file.
}
