% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.build-package.R
\docType{package}
\name{traits.build-package}
\alias{traits.build}
\alias{traits.build-package}
\title{traits.build: Package used to build an AusTraits data resource}
\description{
This package enbales harmonising of data from diverse sources. The code was originally built to support AusTraits, an open-source compilation of data on the traits of Australian plant species. For more information on AusTraits go to https://austraits.org.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://traitecoevo.github.io/traits.build/}
  \item Report bugs at \url{https://github.com/traitecoevo/traits.build/issues}
}

}
\author{
\strong{Maintainer}: Daniel Falster \email{daniel.falster@unsw.edu.au} (\href{https://orcid.org/0000-0002-9814-092X}{ORCID})

Authors:
\itemize{
  \item Elizabeth Wenk (\href{https://orcid.org/0000-0001-5640-5910}{ORCID}) [curator]
}

Other contributors:
\itemize{
  \item ARDC [funder]
  \item ARC [funder]
}

}
\keyword{internal}
